final changes made on various templates

+++ app.py
@@ -1,8 +1,10 @@
-from operator import truth
-from flask import Flask,render_template,Response,jsonify,redirect,request,flash,session,url_for
+from pickle import TRUE
+from flask import Flask,render_template,redirect,request,flash,session,url_for
 from flask_pymongo import PyMongo,ObjectId
 from flask_session import Session
 import bcrypt
+import datetime
+
 
 app= Flask(__name__,static_url_path="/static")
 app.secret_key='Nothing'
@@ -32,8 +34,8 @@ def signup():
             'email': new_email,
             'password': new_password
         })
-        flash(new_username,"Successfully Signed In")
-        return redirect("/")
+        flash(new_username +"Successfully Signed Up")
+        return redirect("/login")
 
     return render_template("signup.html")
 
@@ -45,15 +47,16 @@ def login():
         username=request.form['name']
         session['username']=username
         password=request.form['password']
-        if dbm.find_one({'name':username})  and dbm.find({'password':password}):
-             if username == 'admin':
-                flash(username+" Successfully Signed In")
-                return redirect("/admin")
-             else:
-                flash(username+ " Successfully Signed In")
-                return redirect("/employee")
+        if list(dbm.find({'name':username, 'password':password},{'_id':0,'email':0})):
+            if username == 'admin':
+             flash(username+" Successfully Signed In")
+             return redirect("/admin")
+            else:
+                flash(username + " Successfully Signed In")
+                return redirect("/employee") 
         else:
-            return redirect("/signup")
+            flash('Incorrect Username or Password.Please check if you have signed Up')
+            return redirect("/login")
     return render_template('login.html')
 
 #ADMIN HOME PAGE
@@ -62,8 +65,9 @@ def admin():
     detail=[]
     for i in db.find({},{"_id":0}):
         detail.append(i)
+    newdetail=sorted(detail,key=lambda i:i['ename'])
     # app.add_url_rule("/","/edit")
-    return render_template("admin.html",newuser=detail)
+    return render_template("admin.html",newuser=newdetail)
 
 #EMPLOYEE HOME PAGE
 @app.route("/employee",methods=['GET','POST'])
@@ -87,7 +91,11 @@ def aedit():
         dob=request.form['dob']
         Dept=request.form['Dept']
         LeaveReq=request.form['LeaveReq']
-        adder=db.insert_one({
+        if list(db.find({'eid':eid},{"_id":0,"ename":0,"email":0,'DoB':0,'Dept':0,'LeaveReq':0,'FromDate':0,'NoDays':0,'RoF':0,'ToDate':0,'Ain':0,'Aout':0})):
+            flash('Employee id already exists.Please enter a unique ID')
+            return redirect("/admin/add")
+        else:    
+            adder=db.insert_one({
             'eid': eid ,
             'ename':ename,
             'email':email,
@@ -107,7 +115,6 @@ def eedit():
         empdata=db.find({'ename':username},{'_id':0})
         for i in empdata:
             d.append(i)
-            print(d)
     if request.method == 'POST':
         email=request.form['email']
         dob=request.form['dob']
@@ -130,6 +137,8 @@ def eleave():
     if request.method== 'POST':
         #Nod=request.form['nod']
         FromDate=request.form['FromDate']
+        DFromDate = datetime.datetime.strptime(FromDate, '%Y-%m-%d')
+        # print(type(DFromDate))
         ToDate=request.form['ToDate']
         Nod=request.form['nod']
         Leave=request.form['LeaveReq']
@@ -143,9 +152,8 @@ def eleave():
         }},True)
         return redirect("/employee")
     return render_template("eleave.html")
-        
 
-# EMPLOYEE ATTENDANCE
+
 
 # ADMIN LEAVE REQUESTS PAGE
 @app.route("/admin/leave",methods=['GET','POST'])
@@ -154,6 +162,24 @@ def lq():
     for i in db.find({'LeaveReq': 'Requested'},{"_id":0}):
         ltreq.append(i)
     return render_template('AdminLeaveReq.html',data=ltreq) 
+
+@app.route("/isubmit",methods=['GET','POST'])
+def isub():
+    session['username']=username
+    x=datetime.datetime.now().strftime("%d-%m-%Y %H:%M:%S")
+    print("Ain is "+x)
+    updateain= db.update_one({'ename':username},{'$set':{'Ain':x}})
+    
+    return redirect("/employee")
+
+@app.route("/osubmit",methods=['GET','POST'])
+def osub():
+    session['username']=username
+    x=datetime.datetime.now().strftime("%d-%m-%Y %H:%M:%S")
+    print ("Aout"+ x)
+    updateain= db.update_one({'ename':username},{'$set':{'Aout':x}})
+    return redirect("/employee")
+
 
 
 #ADMIN LEAVE REQUEST ACCPETING/DECLININ EMPLOYEE
@@ -168,10 +194,19 @@ def lsq(id):
 #DELETE FROM ADMIN SIDE
 @app.route("/admin/<id>",methods=['GET','POST'])
 def delete(id):
-    db.delete_one({'eid':id})
-    dbm.delete_one({'eid':id})
+    print(id)
+    x=list(db.find({'eid':id},{'ename':1,'_id':0}))
+    # print(x)
+    y=x[0]['ename']
+    print(y)
+    if dbm.find({'name':y}):
+        print('Got in')
+        db.delete_one({'eid':id})
+        dbm.delete_one({'name':y})
     return redirect(url_for('admin'))
     #return render_template("admin.html")
+    # d1=db.find({'eid':id},{"_id":0,'ename':1})
+    # print(list(d1))
 
 #LOGOUT OF USERS
 @app.route("/logout",methods=['GET','POST'])
